Extract validator logic into a trait.
Having these methods in a trait will let us not copy + paste code into
the elastic search plugin. It could also be useful for other datasource
plugins that need validation features.
